requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
newspaper3k==0.2.8
nltk==3.5
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.56.0
urllib3==1.26.3
Werkzeug==1.0.1
